fix: move printer service and add local list updating on every function that needs it
